accomodates failing failing versions without blowing stuff away.

+++ fetchimages/bulkpngconvert.sh
@@ -12,12 +12,16 @@ for i in $(find . -maxdepth 1 -type d); do
         for tiffile in $(ls *.tif); do
             pngfile=${tiffile:0:8}".png"
             convert -verbose -quality 9 $tiffile $pngfile # Sometimes png is bigger than tif.
-            sizetiff=$(du -k "$tiffile" | cut -f 1)
-            sizepng=$(du -k "$pngfile" | cut -f 1)
-            if [ $sizetiff -lt $sizepng ]; then
-                rm -f $pngfile
+            if [ $? -eq 0 ]; then
+                sizetiff=$(du -k "$tiffile" | cut -f 1)
+                sizepng=$(du -k "$pngfile" | cut -f 1)
+                if [ $sizetiff -lt $sizepng ]; then
+                    rm -f $pngfile
+                else
+                    rm -f $tiffile
+                fi
             else
-                rm -f $tiffile
+                echo "converting failed" > $tiffile".error"
             fi
         done
         cd $START_DIR
